develop
Amélioration du centre d'un noeud. Suppression des arc qui se créaient quand ça ne touchait pas les noeuds
