+++ app/src/main/res/layout/activity_main.xml
@@ -11,7 +11,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     tools:context="fr.istic.graphes.MainActivity">
 
-    <fr.istic.graphes.component.GraphVue
+    <fr.istic.graphes.view.GraphVue
         android:id="@+id/canvas"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
